Adding keyword directive for special meaning such as keyword in C(if, for, ...).

+++ tools/inc/lexer.h
@@ -20,7 +20,7 @@ typedef struct {
 
 lexer_t* new_lexer(int);
 void del_lexer(lexer_t*);
-int advance_token(lexer_t*);
+long advance_token(lexer_t*);
 int peek_token(lexer_t*);
 bool in_first(lexer_t*, ...);
 int is_final_state(int);
